+++ src/main/java/com/packsendme/microservice/sa/businessrule/consumer/component/AirwayBREConsumer_Component.java
@@ -1,18 +1,18 @@
-package com.packsendme.microservice.sa.businessrule.component;
+package com.packsendme.microservice.sa.businessrule.consumer.component;
 
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
 
 @Component
-public class ExecutionBREConsumer_Component implements BRE_Consumer {
+public class AirwayBREConsumer_Component implements BRE_Consumer{
 
 	private String msg;
 	
-	@KafkaListener(topics = "${kafka.topic.executionBRE}")
+	@KafkaListener(topics = "${kafka.topic.airwayBRE_SA_Instance}")
 	public void receive(String data) {
 		this.msg = data;
 		System.out.println(" ------------------------------- ");
-		System.out.println(" topic_executionBRE_sa "+ data);
+		System.out.println(" topic_airway_sa "+ data);
 		System.out.println(" ------------------------------- ");
 	}
 	
@@ -20,4 +20,5 @@ public class ExecutionBREConsumer_Component implements BRE_Consumer {
 		return msg;
 	}
 
+ 
 }
