+++ src/main/java/com/packsendme/microservice/sa/businessrule/consumer/component/BRE_Consumer.java
@@ -1,4 +1,4 @@
-package com.packsendme.microservice.sa.businessrule.component;
+package com.packsendme.microservice.sa.businessrule.consumer.component;
 
 import org.springframework.kafka.annotation.KafkaListener;
 
@@ -6,6 +6,7 @@ public interface BRE_Consumer {
 	
 	@KafkaListener
 	public void receive(String data);
+	
 	
 	public String consumerTopic();
