+++ src/main/java/com/packsendme/microservice/sa/businessrule/consumer/component/ExecutionBREConsumer_Component.java
@@ -1,18 +1,18 @@
-	package com.packsendme.microservice.sa.businessrule.component;
+package com.packsendme.microservice.sa.businessrule.consumer.component;
 
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
 
 @Component
-public class MaritimewayBREConsumer_Component implements BRE_Consumer {
+public class ExecutionBREConsumer_Component implements BRE_Consumer {
 
 	private String msg;
 	
-	@KafkaListener(topics = "${kafka.topic.maritimewayBRE_SA_Instance}")
+	@KafkaListener(topics = "${kafka.topic.executionBRE}")
 	public void receive(String data) {
 		this.msg = data;
 		System.out.println(" ------------------------------- ");
-		System.out.println(" topic_maritimeway_sa "+ data);
+		System.out.println(" topic_executionBRE_sa "+ data);
 		System.out.println(" ------------------------------- ");
 	}
