--- src/main/java/com/packsendme/microservice/sa/businessrule/produce/component/AirwayBREConsumer_Component.java
@@ -1,23 +1,0 @@
-package com.packsendme.microservice.sa.businessrule.component;
-
-import org.springframework.kafka.annotation.KafkaListener;
-import org.springframework.stereotype.Component;
-
-@Component
-public class AirwayBREConsumer_Component implements BRE_Consumer{
-
-	private String msg;
-	
-	@KafkaListener(topics = "${kafka.topic.airwayBRE_SA_Instance}")
-	public void receive(String data) {
-		this.msg = data;
-		System.out.println(" ------------------------------- ");
-		System.out.println(" topic_airway_sa "+ data);
-		System.out.println(" ------------------------------- ");
-	}
-	
-	public String consumerTopic(){
-		return msg;
-	}
-
-}
